Add action for building samples/bpf
We have a bunch of BPF sample code that depends on core Linux and BPF
APIs. These APIs may change over time, but given that these samples are
not included in the 'all' target but have to be built separately, they
are more likely to go stale [0].
To at least ensure a minimum level of up-to-dateness, let's build these
samples as part of the CI. This change introduces an action that we can
use to perform such a build.

[0] https://lore.kernel.org/bpf/BN6PR11MB16331BEF1C7F6F37F76C55DD92899@BN6PR11MB1633.namprd11.prod.outlook.com/T/#mad6fae31be210e5231d63ebffcad305b9a0c7bae

Signed-off-by: Daniel Müller <deso@posteo.net>