Add action for building selftests
This change introduces a dedicated action for building selftests.
Similar to other actions, it can be invoked from GitHub workflows from
other repositories. The script is mostly a copy of a script of the same
name in the kernel-patches/vmtest repository [0] with minor adjustments
(which will be removed once we switch over to consuming this new
action).

[0] https://github.com/kernel-patches/vmtest/blob/2d0290bb10430bf127e823fcdd9eb10bbc5b1424/.github/actions/vmtest/build_selftests.sh

Signed-off-by: Daniel Müller <deso@posteo.net>