Honor and rely on KBUILD_OUTPUT
If we ever want to get incremental kernel build support rolled out to
CI, we need a way to separate build artifacts from source code. That is
necessary because source code is checked out by GitHub Actions as part
of the CI run and it would be counter-productive to keep it around and
transferred between builds.
But even if we decide against incremental builds long-term, it is not a
bad practice to separate those two concerns -- it makes for a logical
separation and can act as a forcing factor to be more explicit about the
paths to artifacts and binaries.
With that in mind, this change adjusts our logic to honor
KBUILD_OUTPUT [0], which controls where the kernel build system stores its
build artifacts. Several of our actions now accept it as input parameter
and the underlying scripts make sure that the build system sees this
variable.

[0]: https://www.kernel.org/doc/html/latest/kbuild/kbuild.html#kbuild-output

Signed-off-by: Daniel Müller <deso@posteo.net>