Retry builds on failure
With incremental building of the kernel in CI being attempted, we need a
way to deal with bugs in the incremental build logic and that could
cause spurious build failures.
With this change we attempt a `make clean; make all`, causing a full
rebuild in case the build failed. That may cause a more or less
unnecessary full rebuild in case a broken patch is submitted, but that's
a relatively rare occurrence and I don't think there is a "proper" way
to detect that. It will also mean that libbpf (which is unlikely to gain
incremental building capabilities anytime soon) may end up doing an
unnecessary rebuild. But again, a spurious build failure there is even
less likely, because we pull upstream and don't have arbitrary
user-provided kernel patches to deal with. On the upside, this retry
logic is very simple and introduces barely any complexity (as opposed
to, say, plumbing through a flag indicating whether it's an incremental
build to begin with) -- and I think that the simplicity arguments wins.
We could also move the retry logic higher up (say, into the workflow
level). But it's not even clear whether that is possible, let alone
anywhere nearly as simple as this change.

Signed-off-by: Daniel Müller <deso@posteo.net>