Rely on upstream kernel configuration
So far we have relied on the kernel configuration as checked into the
kernel-patches/vmtest repository. However, this configuration is now
included in upstream Linux [0].
With this change we switch over to using the configuration from there.
This patch is a respin of 2e4549ba7cf79 ("Rely on upstream kernel
configuration") that considers the possibility of the repository under
test not having the upstreamed kernel configuration available yet, which
is currently the case for the bpf tree (as a sync between bpf-next and
bpf has not yet happened).

[0] https://lore.kernel.org/bpf/165893461358.29339.11641967418379627671.git-patchwork-notify@kernel.org/T/#m2a97b0ea9ef0ddee7a53bbf7919e3f324b233937

Signed-off-by: Daniel Müller <deso@posteo.net>