Introduce get-llvm-version action
Whenever development on a new LLVM version starts, earlier version of
llvm, clang, and lld may seize to exist in distribution snapshot
repositories. For us, that means having to manually bump the version in
three repositories [0][1][2], in a specific order. That's cumbersome to
do and ultimately we probably want a single source of truth as to what
the most recent LLVM version to use is.
This change introduces an action that centralizes this information while
making it available across repository boundaries. We will subsequently
be able to eliminate hardcoded version information in the other two
repositories constituting the BPF CI.

[0] https://github.com/kernel-patches/vmtest/pull/107
[1] https://github.com/libbpf/libbpf/pull/542
[2] https://github.com/libbpf/ci/pull/24