Make Qemu image path configurable
The path to the file system image used by our invocation of Qemu is
currently hard coded to /tmp/root.img. This constant originates in the
run_vmtest.sh script, but is then hard-coded in other locations as well,
without an obvious connection between the two. That makes it hard to
understand the code.
This change makes the path to the image an explicit input parameter to
the prepare-rootfs action and the run_vmtest.sh script, which will make
it possible to make the input-output relationship more obvious in client
code moving forward.

Signed-off-by: Daniel Müller <deso@posteo.net>