Plumb through test to run
We would like to have the ability to run selftests in parallel as
opposed to fully serial as is currently the case. In order to make that
possible, this change adjusts the prepare-rootfs action to accept an
optional argument specifying the test to run and passes it to the
run_selftests.sh script.

Signed-off-by: Daniel Müller <deso@posteo.net>