Move bpftool check into run-qemu action
Currently the bpftool tests are run as part of the prepare-rootfs
action. That's counter-intuitive, as running a test has nothing
"preparational" to it. That in turn causes confusion with users that are
trying to debug failures of said tests: they have no idea how to find
the logs in question, because success/failure is reported as part of a
later action.
This change moves the running of the bpftool tests into the run-qemu
action, which is where selftest results are reported as well. As such,
we remove the need for an implicit file with a hard coded path being
used for communication between these two steps, and of course the
confusion for users is fixed as well.
It is still sub-optimal because we are running this test for every other
selftest run, but that is unchanged from before. The remaining issues
will be addressed at a later point.

Signed-off-by: Daniel Müller <deso@posteo.net>