Remove default arguments from prepare-rootfs action
Having default arguments on an action should probably be avoided if
possible. For one, it's confusing to see something being "required" but
having a default value, but if not explicitly mentioned on the call
site, it's also unnecessarily hard to understand what is going on.

With https://github.com/libbpf/libbpf/pull/570 and
https://github.com/kernel-patches/vmtest/pull/131 merged we no longer
need default values for two arguments to the prepare-rootfs action, as
all call sites explicitly provide values. Hence, remove them.

Signed-off-by: Daniel Müller <deso@posteo.net>