Add kernel-root argument to run-qemu action
This change adds the kernel-root argument to the run-qemu action. We
will need this argument once we want to move bpftool checks into the
actual selftest run job, as opposed to having them happen in the
unrelated prepare-rootfs. We need a default value while all "callers"
are being transitioned to providing this argument. The KERNEL_ROOT
variable that we set based on it is not currently used (will come with
subsequent changes).

Signed-off-by: Daniel Müller <deso@posteo.net>