ci: symlink /dev/fd to /proc/self/fd
For process substitution to work, /dev/fd needs to be symlinked to
/proc/self/fd.

Failing to do so, we get ENOENT when using process substitution:

```
[root@(none) /]# read a < <(echo 1 2 3)
sh: /dev/fd/63: No such file or directory
[root@(none) /]# ln -s /proc/self/fd /dev/fd
[root@(none) /]# read a < <(echo 1 2 3)
[root@(none) /]# echo $a
1 2 3
```

In `/etc/rcS.d/S10-mount` we mount both procfs and devtmpfs. After that,
/dev/fd does not exist.
This patch creates a symlink from /dev/fd to /proc/self/fd if the file
does not exist yet.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>