ci: Update rootfs images to 2022-10-13 version
This is a downgrade from bookworm to bullseye to be able to test verify_pkcs7_sig

verify_pkcs7_sig was failing with the following error:
```
   test_verify_pkcs7_sig:FAIL:populate_data_item_str unexpected error: 127 (errno 22)
```

This indicated that it failed to run a command (127 == command not
found). Further investigation indicated that `sign-file` failed with:
```
 ./sign-file -h
./sign-file: error while loading shared libraries: libcrypto.so.1.1: cannot open shared object file: No such file or directory
```

We compile the kernel/tools in Ubuntu focal, which ships libssl1.1.
Debian bookworm ships libssl3.
By downgrading to Debian bullseye, we have a matching libssl1.1

Signed-off-by: Manu Bretelle <chantr4@gmail.com>