rootfs: Fix rootfs creation on foreign architecture
An update in upstream cause a binary on the foreign platform to be
executed, which in turned caused the rootfs creation to fail with
something like:
```
chroot: failed to run command '/bin/true': Exec format error
```

`debootstrap` can handle that natively by building the image in 2 steps.
1) download the packages, verify, and extract them inside the chroot
2) running a second stage that unpack and configure them.

Before the second stage, we copy `qemu-${foreign_arch}-static` into the
chroot, allowing the native binaries to run properly.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>