[ansible] move s390x ansible setup in its own file
s390x has a specific setup currently. for instance, it needs to create
its own actions-runner docker image, the actions-runner are the ones
from upstream x86_64 version which are ran under qemu-static.
On the other hand, x86_64 and later aarch64 will be able to use a
generic setup that hardly change between the 2.

This diff moves the s390x setup in its own file and includes it only
when the host architecture is s390x. Otherwise, it will include the
generic file (which is work TBD).

Signed-off-by: Manu Bretelle <chantr4@gmail.com>