[ansible] create units/instances based on runner_repo_list values
AWS used to read the repo name from the ec2 instance metadata tags.
This used to be fine when running 1 runner per EC2 instances.

With this change, the repositories/instances are read from the
`runner_repo_list` variable, like for s390x. This allows us to run
multiple instances of the runner for a given repository, as well as
multiple repository runners.
For instance, run 8 kernel-patches/bpf runners and 1
kernel-patches/vmtest runner per hosts.

Do do that, the name of the systemd service unit needs to be changed. We
normalize the name of the repository to create a service unit under
`/etc/systemd/system/actions-runner-{normalized-name}@.service`

Within that service unit, we discover which env file to source by using
the `%p`, this is also used to name the docker container.
More on specifiers at
https://www.freedesktop.org/software/systemd/man/systemd.unit.html#Specifiers
But TL;DR, we use `%p` vs `%P` as the latter would "unescape" and replace
`-` with `/`, making it an invalid container name.
`%i` is the worker number.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>