Handle incremental builds gracefully
We intend to take a stab at supporting incremental kernel builds in CI
with https://github.com/kernel-patches/vmtest/pull/166. To make that
happen we carry over build artifacts stored in KBUILD_OUTPUT. One of
those artifacts is .config. Now so far we have unconditionally created
and used a new configuration. But that may cause unnecessary rebuilds.

With this change we only update the kernel configuration if it really
changed.

Signed-off-by: Daniel Müller <deso@posteo.net>