[ansible] using ansible_hostname instead of inventory_hostname
`ansible_hostname` is the hostnae as returned by `hostname` while
`inventory_hostname` is the name in the inventory file.
In the first iteration of the use of this playbook, we will trigger the
run of ansible locally, which will make all runners have the same name
otherwise.
This change uses the hostname instead which in our current setup are
unique.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>