ansible: Add architecture label
The runner by default adds a label (X64), but this is based on the
architecture the runner was built for, not the host architecture.
For instance, on s390x, it still shows `X64` because we use the X64
binary.

This change adds the achitecture returned by `uname -m` to the label
list. Later, we will be able to use those to filter which hosts to run
on.

```
-LABELS=ubuntu-latest,docker-main
+LABELS=ubuntu-latest,x86_64,docker-main
```

Signed-off-by: Manu Bretelle <chantr4@gmail.com>