Introduce --source argument to prepare-rootfs/run.sh
With upcoming changes we would like to separate kernel build artifacts
from kernel sources. Unfortunately, it turns out the two concerns are
not really separated by prepare-rootfs/run.sh. Specifically, the script
assumes `make -C "${BUILDDIR}" <xxx>` to work. However, the build
directory does not necessarily contain any makefiles at all. What the
script really means to do is `make -C <kernel-source> <xxx>`.
Unfortunately, currently the script does not receive the location of the
kernel source tree to begin with.
To fix this state of affairs, this change introduces the `--source`
option, which can be used to provide the location of the kernel source
tree.

Signed-off-by: Daniel Müller <deso@posteo.net>