ci: limit number of CPUs to 8 for any qemu instance
Limit number of CPUs dedicated to qemu instance to 8, if it happens that
host has more available. This limits excessive parallelism of test_maps
test, which will help prevent random OOMs. For non-parallel runs this
makes no difference. For parallels there might be a bit of slow down,
but 8 is plenty anyways.

Signed-off-by: Andrii Nakryiko <andrii@kernel.org>