ansible: Add a script to generate a temporary ACCESS_TOKEN
`gh_token_generator.sh` takes all the arguments needed to generate a
token and write it into the destination file (the last argument).
One of the reason for the need of the last argument is that systemd
ExecPreStart (and friends) do not suppor file redirection, so we need to
do this within a script.
https://www.freedesktop.org/software/systemd/man/systemd.service.html#Command%20lines

The idea is that the script will generate a new file with environment
variables that later get used by `docker`.
Because we load this file second, any env var it declares that may have
existed in the first file are going to override them.
Specifically, even though ACCESS_TOKEN may be declared in the first
file, the fact that we redeclare it here will override it.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>