Honor patches with .patch extension
This change prepares the patch-kernel action for the application of
patches with the extension .patch, in addition to .diff. The need to
rename patches to sport the .diff extension has repeatedly been shown to
be a difficult thing to remember.
See https://github.com/kernel-patches/vmtest/issues/160 for additional
technical details.

Signed-off-by: Daniel Müller <deso@posteo.net>