ansible: Add support for an optional runner prefix
Currently, runner_names are a combinaiton of instance_id (or hostname
for non-aws hosts) and worker number.
With a mixture of runners running on baremetal and some running on VMs,
it can be hard to differentiate once they all use the ansible playbook
to be provisioned.
Here we had an optional runner_prefix that can be set to hint at what
type of machines we are looking at, e.g one can set "vm" for a vm, "bm"
for a baremetal host.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>