build: Add MAX_BUILD_PROC env variable to control to parallelism of make jobs
When running on bare-metal, we use multiple containers per host,
currently 4*nproc

On a machine with 96 CPUs, we run 12 runners. If all runners are trying
to build something, that gets us to  `4*96*12 = 4608` make processes
trying to build, getting the load of the host to got high and possibly
locking the host down.

The goal of `MAX_MAKE_JOBS` is to be able to limit the number of
processes that are used while building and find a sweet spot between
performance and overcommitment.

```
$ source ./helpers.sh
[chantra@worktop ci]$ MAX_MAKE_JOBS=4 kernel_build_make_jobs
4
[chantra@worktop ci]$ MAX_MAKE_JOBS=64 kernel_build_make_jobs
32
[chantra@worktop ci]$ kernel_build_make_jobs
32
[chantra@worktop ci]$ MAX_BUILD_PROC= kernel_build_make_jobs
32
```
Signed-off-by: Manu Bretelle <chantr4@gmail.com>