build: add `max-make-jobs` input to action.yml files
This is used to control how many jobs make will use.
By default, we have it set to 4*nproc, which is fine when we run
dedicated docker jobs per vms, but when running multiple docker jobs on
a hosts, nproc is no more representative of what is available to the
runner.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>