ansible: put runners's work dir on tmpfs for bare metal hosts
When running on bare metal, when all jobs are busy, we end up doing a
lot of io.
Our runners are ephemeral, we clean the containers when we are done with
a job so we don't care about data retention.
Our bare metal machines also havemore memory available than they
actually use.
This change:
1) use a variable to set the runner workdir
2) mount this directory on tmpfs

This way, all git, tar, zstd, compiling, qemu runs are done on top of
memory backed FS vs EBS.

We need to instruct docker to mount the folder with `exec` on as
depending on the version, it may default to noexec, preventing scripts
from executing.