ansible: remove unbuntu-latest tag from self-hosted runners
This is related to kernel-patches/vmtest#162

ubuntu-latest is a tag used by GH owned runners.
If we still use this tag we may have issues with job being ran on
aarch64 arch when expecting x86_64, and, once runners are becoming
org-level runners, they would be picked up by projects we don't want to
run on self-hosted.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>