Make kbuild-output input required in all actions
At this point, there is no reason to build without KBUILD_OUTPUT
specified: maintaining additional code paths that build that way is just
a burden and there is no benefit of building within the source tree
versus outside of it.
Hence, this change proposes making the kbuild-output argument to the
various actions mandatory. It's already provided by all clients, so
there is no effective change.

Signed-off-by: Daniel Müller <deso@posteo.net>