ansible: Move common components to main.yml
s390x is being streamlined with the rest of the deployments.
Aside from how the container is built, the end goal is that we will use
the same configurations models between x64, arm64 and s390x.

This change move the creation of common files (env var, secrets...) into
main.yml.

When running this on a AWS host, nothing changes. On a s390x host, aside
form newly created files, nothing has changed.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>