ansible: converge s390x runners with the rest of runners
Now, the only thing that s390x role is doing is to:
- sync libbpf/ci
- build the docker image.

All docker jobs will now use the same image (while before they used to
be a docker image per worker flavor).
Once we build the image under kernel-patches/runner, the remainder of
main-s390x.yml will go.

This was pushed to `bpf-ci-runner-s390x-7`. The runners could attach to
the orgnization and are available to the repositories.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>