Print diff when updating kernel config to use
The build-linux action conditionally updates the kernel config to use if
it changed. It's generally useful to be able to see what exactly has
changed. For example, the (expanded) configuration contains the
compiler version and given that we install the newest clang snapshot on
each run we may end up with a new compiler version compared to the one
that created the build artifacts and  that forces a full rebuild. It's
generally non-obvious when that happens and so printing a diff of the
two configs can help debug issues.

Signed-off-by: Daniel Müller <deso@posteo.net>