ansible: Create a default gh_app_id value of 0
When using organization-level runners, we use a GH app.
That app has an ID and cert used to generate an ACCESS_TOKEN.

When generating the runner's .service file, we check if the repo GH App
matches any of the GH apps we have registered in out config. If so, an
extra pre-run script is added to the start sequence.

If a repository does not define the key "gh_app_id", our check would
fail because the key is undefined.
The trick here is that we set a GH app Id by default that does not
exist, hence the check will never match, but it will not assert either.

Long term, we should only support GH app and not PAT, so I am more
inclined in using this as a middle step than properly supporting PAT and
Apps.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>