ansible: Set runner_name_prefix to the host's hostname by default
For Amazon Linux running in EC2, the prefix get overwritten by the
instance ID instead.
We try to do this in a block where failure are ignored.
This code is only run when the host OS is Amazon Linux. Failure to
access the EC2 endpoint to discover the host metadata will result in the
host defaulting to use its hostname. This is useful when running an
Amazon Linux host in a VM for local testing.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>