[run-qemu] consume test_progs json report
This change makes the qemu action consume the report generated by
test_progs (introduced in kernel-patches/vmtest#207) and generates
github annotations [0] of the summary as well as the failing tests.
It also generates a summary of the job [1].

The end goal is to provide an easier way for people to directly access
the tests that failed instead of having to scroll through pages of logs.

[0] https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-error-message
[1] https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#adding-a-job-summary

Signed-off-by: Manu Bretelle <chantr4@gmail.com>