[run-qemu][print_test_summary] build_summaries_for_test to generate individual test summaries
The logic between handling summaries for test and subtest was very
similar.
This function receive a (sub)test dict and generate its textual
summaries for both GH job summary and GH test console.

When asubtest is used, the parent test name and number must be passed as
arguments to the function.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>