[ansible] force aptman/qus with tag d7.1
We have had issue with the latest tag on newly provisioned s390x
machines. More specifically some x86_64 command would fails while some
other were not. The exact failure is yet TBD, but through
troubleshooting and narrowing down the scope of the issue, it appears to have
been a regression with the latest container version of aptman/qus, which is
currently 7.2.
This pins the container to the debian version of the 7.1 release.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>