run-vmtest: consider all deployments rc unless it is kernel-patches/bpf
I can't think of a good way to make this work. But currently, it would make tests fail
when running the action on libbpf/ci.
We need to rethink how we deal with "deployment", but for now, this is essentially considering
all deployments rc (and denylisitng more tests, including for kernel-patches/vmtest), but at least
we can have some coverage of libbpf/ci itself.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>