Reorganize reusable workflows
We want to maintain only one copy of each reusable workflow between
libbpf/ci and kernel-patches/vmtests [1]. To achieve that, libbpf/ci
is going to be the source of truth for them, and kernel-patches/vmtest
main workflow will be calling libbpf/ci.

This change reorganizes files to make it clear which workflows are
reusable, and which are repository-specific. I suggest the following
convention:

* `.github` contains workflows local to the owner repository
* `workflows` contains reusable workflows, intened for external use

[1]: https://github.com/kernel-patches/vmtest/tree/master/.github