build-selftests: v2 of the action inputs
Changes in action inputs:
  * kernel-root is now a required input
  * kernel, kbuild-output and max-make-jobs are removed

kernel-root input is required, because to build selftests
one must to clone a kernel source tree first.

The build script behavior can be directed more granularly with env
variables:
  * KBUILD_OUTPUT now defaults to the kernel-root
  * VMLINUX_BTF now can be set externally
  * MAX_MAKE_JOBS can be set as env variable if desired

Preparation script logic is removed: the rationale is that the caller
is responsible for setting up required environment before building.

Signed-off-by: Ihor Solodrai <ihor.solodrai@pm.me>