[ansible][qemu-user-static] Force automount to mount /proc/sys/fs/binfmt_misc
systemd is kind of taking over that directory. While Ubuntu seems to
handle this just right, Amazon Linux 2 does not, essentially, the
directory is handled by automount, but is not mounted and onl;y ends up
mounted once we try to access it with `ls` or such.
This gets us to end up in a situation where qemu-user-static has ran,
but its config gets overridden.

I did not manage to get this to work properly on AL2, but we can trick
it by "pre-warming" the mount by running a `ls /proc/sys/fs/binfmt_misc`
as a pre-exec-start step (ignoring failures).

Tried this on Ubuntu host, and there were still happy.
Signed-off-by: Manu Bretelle <chantr4@gmail.com>