kernel-build: build GCC BPF test runner on x86_64
When BUILD_BPF_GCC is set in kernel-build, GCC BPF will be built and
BPF_GCC variable set for build-selftests. When BPF_GCC is set,
test_progs-bpf_gcc runner is built in selftests/bpf.

If test_progs-bpf_gcc is specified in KERNEL_TEST, it will be executed
by run-vmtest action.

Signed-off-by: Ihor Solodrai <ihor.solodrai@pm.me>