GCC BPF: factor out into a separate job
In order to test selftests/bpf build with GCC BPF, the latest snapshot
of GCC BPF compiler has to built from source.

We don't want this to be a dependency for all the test runners, so
introduce a separate gcc-bpf workflow that executes the following:
  * checkout Linux source tree and patch it (as in kernel-build)
  * download relevant artifacts with the KBUILD_OUTPUT objects (as in
    kernel-test)
  * build/download GCC BPF compiler
  * build selftests/bpf/test_progs-bpf_gcc runner, hence the test BPF
    objects built with GCC

The test_progs-bpf_gcc runner is not executed on BPF CI yet [1], so
successful selftests build is a successful job run.

[1] https://lore.kernel.org/bpf/87bjw6qpje.fsf@oracle.com/