run-vmtest: add path normalization script
actions/upload-artifacts has restrictions on what characters can be
used in uploaded paths.

Add a script that replaces forbidden characters in all paths under a
specified directory.

We ran into this problem when trying to upload traffic monitor
logs (which may contain a colon `:`). But since it's a github-specific
restriction, it doesn't make sense to enforce it upstream.

Signed-off-by: Ihor Solodrai <ihor.solodrai@linux.dev>