build-linux: always overwrite kconfig
In the main kernel build script there is a logic of comparing clean
kconfig with cached kconfig, and then updating it if necessary.

This is unnecessary, and in some cases may cause issues if the cached
config is bogus.

Always overwrite the config with the one generated from a specified
list of config files.

Signed-off-by: Ihor Solodrai <ihor.solodrai@linux.dev>