run-vmtest: exit gracefully from print_test_summary.py
When selftests fail because of a watchdog timeout or segfault or
anything similar, a JSON summary might be empty or absent.

Currently this leads to print_test_summary.py to verbosly fail with an
exception, which is confusing when looking at the job log.

Handle common errors gracefully in the script.

Signed-off-by: Ihor Solodrai <isolodrai@meta.com>