run-vmtest: Allow specifying cpu and memory for vms
The max-cpu parameter is currently unused by the script and therefore
vmtest defaults to 2 CPU. There is not parameter available to specify
the amount of memory to give to the vm, for this the vmtest defaults to
4G.

These low defaults(specifically the memory) sometimes cause the bpf
seltests to sporadically fail. So, add the option to specify memory and
also start using the max-cpu parameter with vmtest.

Signed-off-by: Puranjay Mohan <puranjay@kernel.org>